fixed minishell + many other errors

+++ parsing/find_parts_num.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   find_pipes_num.c                                   :+:      :+:    :+:   */
+/*   find_parts_num.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alsaeed <alsaeed@student.42abudhabi.ae>    +#+  +:+       +#+        */
+/*   By: alsaeed <alsaeed@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/16 21:45:33 by alsaeed           #+#    #+#             */
-/*   Updated: 2023/12/25 22:03:57 by alsaeed          ###   ########.fr       */
+/*   Updated: 2024/01/04 20:59:55 by alsaeed          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	find_parts_num(char *str)
 
 	i = -1;
 	parts_num = 1;
-	while (++i < ft_strlen(str))
+	while (++i < (int)ft_strlen(str))
 	{
 		if (str[i] == '|')
 			parts_num++;
